--- Entity/subscriptionwall.php
@@ -1,313 +1,0 @@
-<?php
-
-/**
- * Auto generated by MySQL Workbench Schema Exporter.
- * Version 3.0.2 (doctrine2-annotation) on 2016-04-22 11:53:36.
- * Goto https://github.com/johmue/mysql-workbench-schema-exporter for more
- * information.
- */
-
-namespace Integrated\Bundle\SubscriptionBundle\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-use Doctrine\Common\Collections\ArrayCollection;
-
-/**
- * Entity\SubscriptionWall
- *
- * @ORM\Entity(repositoryClass="SubscriptionWallRepository")
- * @ORM\Table(name="subscription_wall")
- */
-class SubscriptionWall
-{
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="UUID")
-     * @ORM\Column(type="string", length=36)
-     */
-    protected $id;
-
-    /**
-     * @ORM\Column(name="`name`", type="string", length=200)
-     */
-    protected $name;
-
-    /**
-     * @ORM\Column(type="string", length=1000, nullable=true)
-     */
-    protected $teaser;
-
-    /**
-     * @ORM\Column(type="boolean", nullable=true)
-     */
-    protected $disabled;
-
-    /**
-     * @ORM\Column(type="string", length=45, nullable=true)
-     */
-    protected $freetier;
-
-    /**
-     * @ORM\Column(type="string", length=45, nullable=true)
-     */
-    protected $channels;
-
-    /**
-     * @ORM\Column(name="`values`", type="string", length=45, nullable=true)
-     */
-    protected $values;
-
-    /**
-     * @ORM\ManyToMany(targetEntity="SubscriptionType", mappedBy="subscriptionWalls")
-     */
-    protected $subscriptionTypes;
-
-    /**
-     * @ORM\ManyToMany(targetEntity="ContentChannel", mappedBy="subscriptionWalls")
-     */
-    protected $contentChannels;
-
-    public function __construct()
-    {
-        $this->subscriptionTypes = new ArrayCollection();
-        $this->contentChannels = new ArrayCollection();
-    }
-
-    /**
-     * Set the value of id.
-     *
-     * @param string $id
-     * @return \Entity\SubscriptionWall
-     */
-    public function setId($id)
-    {
-        $this->id = $id;
-
-        return $this;
-    }
-
-    /**
-     * Get the value of id.
-     *
-     * @return string
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * Set the value of name.
-     *
-     * @param string $name
-     * @return \Entity\SubscriptionWall
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    /**
-     * Get the value of name.
-     *
-     * @return string
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * Set the value of teaser.
-     *
-     * @param string $teaser
-     * @return \Entity\SubscriptionWall
-     */
-    public function setTeaser($teaser)
-    {
-        $this->teaser = $teaser;
-
-        return $this;
-    }
-
-    /**
-     * Get the value of teaser.
-     *
-     * @return string
-     */
-    public function getTeaser()
-    {
-        return $this->teaser;
-    }
-
-    /**
-     * Set the value of disabled.
-     *
-     * @param boolean $disabled
-     * @return \Entity\SubscriptionWall
-     */
-    public function setDisabled($disabled)
-    {
-        $this->disabled = $disabled;
-
-        return $this;
-    }
-
-    /**
-     * Get the value of disabled.
-     *
-     * @return boolean
-     */
-    public function getDisabled()
-    {
-        return $this->disabled;
-    }
-
-    /**
-     * Set the value of freetier.
-     *
-     * @param string $freetier
-     * @return \Entity\SubscriptionWall
-     */
-    public function setFreetier($freetier)
-    {
-        $this->freetier = $freetier;
-
-        return $this;
-    }
-
-    /**
-     * Get the value of freetier.
-     *
-     * @return string
-     */
-    public function getFreetier()
-    {
-        return $this->freetier;
-    }
-
-    /**
-     * Set the value of channels.
-     *
-     * @param string $channels
-     * @return \Entity\SubscriptionWall
-     */
-    public function setChannels($channels)
-    {
-        $this->channels = $channels;
-
-        return $this;
-    }
-
-    /**
-     * Get the value of channels.
-     *
-     * @return string
-     */
-    public function getChannels()
-    {
-        return $this->channels;
-    }
-
-    /**
-     * Set the value of values.
-     *
-     * @param string $values
-     * @return \Entity\SubscriptionWall
-     */
-    public function setValues($values)
-    {
-        $this->values = $values;
-
-        return $this;
-    }
-
-    /**
-     * Get the value of values.
-     *
-     * @return string
-     */
-    public function getValues()
-    {
-        return $this->values;
-    }
-
-    /**
-     * Add SubscriptionType entity to collection.
-     *
-     * @param \Entity\SubscriptionType $subscriptionType
-     * @return \Entity\SubscriptionWall
-     */
-    public function addSubscriptionType(SubscriptionType $subscriptionType)
-    {
-        $this->subscriptionTypes[] = $subscriptionType;
-
-        return $this;
-    }
-
-    /**
-     * Remove SubscriptionType entity from collection.
-     *
-     * @param \Entity\SubscriptionType $subscriptionType
-     * @return \Entity\SubscriptionWall
-     */
-    public function removeSubscriptionType(SubscriptionType $subscriptionType)
-    {
-        $this->subscriptionTypes->removeElement($subscriptionType);
-
-        return $this;
-    }
-
-    /**
-     * Get SubscriptionType entity collection.
-     *
-     * @return \Doctrine\Common\Collections\Collection
-     */
-    public function getSubscriptionTypes()
-    {
-        return $this->subscriptionTypes;
-    }
-
-    /**
-     * Add ContentChannel entity to collection.
-     *
-     * @param \Entity\ContentChannel $contentChannel
-     * @return \Entity\SubscriptionWall
-     */
-    public function addContentChannel(ContentChannel $contentChannel)
-    {
-        $this->contentChannels[] = $contentChannel;
-
-        return $this;
-    }
-
-    /**
-     * Remove ContentChannel entity from collection.
-     *
-     * @param \Entity\ContentChannel $contentChannel
-     * @return \Entity\SubscriptionWall
-     */
-    public function removeContentChannel(ContentChannel $contentChannel)
-    {
-        $this->contentChannels->removeElement($contentChannel);
-
-        return $this;
-    }
-
-    /**
-     * Get ContentChannel entity collection.
-     *
-     * @return \Doctrine\Common\Collections\Collection
-     */
-    public function getContentChannels()
-    {
-        return $this->contentChannels;
-    }
-
-    public function __sleep()
-    {
-        return array('id', 'name', 'teaser', 'disabled', 'freetier', 'channels', 'values');
-    }
-}
